COL-515 Remove bookmark link from viewer
This setting already exists in the Mirador setting.
The setting that needs to be set to "false" is "mainMenuSettings.buttons.bookmark", as stated in the Github documentation. The dots between the names indicate that the names are "objects", as shown below:

'mainMenuSettings': {
      'show': true,
      'buttons' : {
        'bookmark' : true,
        'layout' : true,
        'options' : false,
        'fullScreenViewer': true
      }
      //'height': 25,
      //'width': '100%'
    }

The above is the default settings found in settings.js in the uncompiled Mirador source code.