Add jquery migrate to match existing jquery to 1.9.1 introduced by Mirador
Mirador includes jQuery and jQuery by default, 1.9.1 and 1.9.2.
Paleography uses jQuery 1.8.7.
Specifically the $.browser object is removed for 1.9, and is used by Drupal overlay. This causes the "manage" tab to not appear.
The ideal is to removed Mirador's jQuery libraries and let it use Drupal's, but we don't know if Mirador works with 1.8.
This resolves COL-649